Feat added Navbar Score
