+++ client/gp-p2/src/pages/HomePage.jsx
@@ -2,26 +2,10 @@ import { useNavigate } from "react-router-dom";
 import background from "../assets/Background_Space.webp";
 import rsp from "../assets/RSP.png";
 import Navbar from "../components/Navbar";
-import axios from "axios";
-import { useEffect, useState } from "react"; //
 
 const HomePage = () => {
-  const navigate = useNavigate();
-  const id = localStorage.getItem("id");
-  const [user, setUser] = useState("");
-
-  const getUser = async () => {
-    try {
-      const { data } = await axios.get("http://localhost:3000/users", id);
-      setUser(data);
-    } catch (err) {
-      console.log(err);
-    }
-  };
-
-  useEffect(() => {
-    getUser();
-  }, []);
+  const navigate = useNavigate()
+  
 
   const play = () => {
     navigate("/battle");
